feat: search

+++ src/renderer/copy.cjs
@@ -4,5 +4,7 @@ const fs = require('fs-extra')
 const ROOT = path.resolve(__dirname, '../../')
 
 const renderer = path.join(ROOT, 'dist/renderer')
-fs.removeSync(renderer)
+if (fs.existsSync(renderer)) {
+  fs.removeSync(renderer)
+}
 fs.moveSync('./dist', renderer)
